Game_Site/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-allauth==0.51.0
django-bootstrap-forms==0.1
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.21.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.12
